requirements.txt
pandas
fastapi
uvicorn
beautifulsoup4
requests
ftfy
urllib3
sentence-transformers
faiss-cpu
streamlit
litellm
python-dotenv
